fix lint and platform check

+++ tests/test_posix_linux.py
@@ -32,14 +32,16 @@ class LinuxPosixTests(unittest.TestCase):
         """
         set up the test case
         """
-        self.assertTrue(os.path.exists(self._shm_location), "shm location does not exist")
-        self._shm_name = str(time.time())
+        if sys.platform.startswith("linux"):
+            # there is one test to be executed outside linux so we need this special check
+            self.assertTrue(os.path.exists(self._shm_location), "shm location does not exist")
+            self._shm_name = str(time.time())
 
-        l = shmlock.ShmLock(self._shm_name)
-        with l:
-            # file should be generated at desired location
-            self.assertTrue(os.path.isfile(os.path.join(self._shm_location,
-                                                        self._shm_name)))
+            l = shmlock.ShmLock(self._shm_name)
+            with l:
+                # file should be generated at desired location
+                self.assertTrue(os.path.isfile(os.path.join(self._shm_location,
+                                                            self._shm_name)))
 
     @unittest.skipUnless(sys.platform.startswith("linux"), "test only for linux")
     def test_empty_shared_memory_file(self):
@@ -51,7 +53,7 @@ class LinuxPosixTests(unittest.TestCase):
 
         # create empty file to fake flawed shared memory file to which shared memory cannot attach
         with open(os.path.join(self._shm_location,
-                               self._shm_name), "w+") as f:
+                               self._shm_name), "w+", encoding="utf-8") as _:
             pass
 
         with self.assertRaises(shmlock.shmlock_exceptions.ShmLockValueError):
@@ -64,6 +66,7 @@ class LinuxPosixTests(unittest.TestCase):
 
         # fake creation of block but lock did not write its uuid to the file. this happens
         # if the process is interrupted right after creation of shared memory file
+        shm = None
         try:
             shm = shared_memory.SharedMemory(name=self._shm_name,
                                             create=True,
@@ -75,8 +78,9 @@ class LinuxPosixTests(unittest.TestCase):
                 l.query_for_error_after_interrupt()
 
         finally:
-            shm.close()
-            shm.unlink()
+            if shm is not None:
+                shm.close()
+                shm.unlink()
 
 
     @unittest.skipUnless(sys.platform.startswith("linux"), "test only for linux")
